pyproject.toml
## ============================================================
## PYPROJECT.TOML - Modern Python package configuration
## ============================================================
## This file tells pip how to install the package and what
## dependencies it needs. Replaces the old setup.py approach.
## ============================================================

[build-system]
## Build tools needed to create the package
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
## Package metadata
name = "digger"
version = "0.1.0"
description = "Foul-mouthed Aussie study assistant with RAG and voice"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "George Wu", email = "george@example.com"}
]

## Dependencies - packages that must be installed
dependencies = [
    "requests>=2.28.0",      ## HTTP requests for ElevenLabs API
    "python-dotenv>=1.0.0",  ## Load .env files for secrets
    "pyyaml>=6.0",           ## Config file parsing
    "rich>=13.0.0",          ## Pretty terminal output
]

## Optional dependencies for development
[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
]

## CLI entry point - makes 'digger' command available after install
## Format: command_name = "package.module:function"
[project.scripts]
digger = "digger.cli:main"

[tool.setuptools.packages.find]
## Tell setuptools where to find Python packages
where = ["."]
